fee prototype

+++ app/routes/category_routes.js
@@ -5,23 +5,27 @@ import authorize from "../middlewares/middleware.js";
 
 const router = Router();
 
-router.post("/category", authorize(Permission.ADD_CATEGORIES), async (req, res) => {
+router.post("/category", async (req, res) => {
   const name = req.body;
 
   if (!req.body.name) {
     res.status(400).json({ message: "Name is required" });
   } else {
-    const name_exist = await prisma.category.findFirst({ where: { name: req.body.name } });
+    const name_exist = await prisma.category.findFirst({
+      where: { name: req.body.name },
+    });
     if (name_exist) {
       res.status(400).json({ message: "Name is already" });
     } else {
       const category = await prisma.category.create({ data: name });
-      res.status(200).json({ message: "Category created successfully", category });
+      res
+        .status(200)
+        .json({ message: "Category created successfully", category });
     }
   }
 });
 
-router.get("/category", authorize(Permission.BROWSE_CATEGORIES), async (req, res) => {
+router.get("/category", async (req, res) => {
   const category = await prisma.category.findMany();
   if (category.length === 0) {
     res.status(404).json({ message: "Category is Empty" });
@@ -30,15 +34,34 @@ router.get("/category", authorize(Permission.BROWSE_CATEGORIES), async (req, res
   }
 });
 
-router.put("/category/:id", authorize(Permission.EDIT_CATEGORIES), async (req, res) => {
+router.get("/category/:id", async (req, res) => {
+  if (isNaN(req.params.id)) {
+    return res.status(400).json({ message: "Invalid ID" });
+  }
+  const category_id = +req.params.id;
+  const category = await prisma.category.findFirst({
+    where: { id: category_id },
+  })
+  if(!category) {
+    res.status(404).json({ message: "Category not found" });
+  } else {
+    res.status(200).json(category);
+  }
+});
+
+router.put("/category/:id", async (req, res) => {
   if (isNaN(req.params.id)) {
     res.status(400).json({ message: "Invalid ID" });
   } else {
-    const category_id = await prisma.category.findFirst({ where: { id: Number(req.params.id) } });
+    const category_id = await prisma.category.findFirst({
+      where: { id: Number(req.params.id) },
+    });
     if (!category_id) {
       res.status(404).json({ message: "Category not found" });
     } else {
-      const category_exist = await prisma.category.findFirst({ where: { name: req.body.name } });
+      const category_exist = await prisma.category.findFirst({
+        where: { name: req.body.name },
+      });
       if (category_exist) {
         res.status(400).json({ message: "Category Name is already" });
       } else {
@@ -52,19 +75,25 @@ router.put("/category/:id", authorize(Permission.EDIT_CATEGORIES), async (req, r
   }
 });
 
-router.delete("/category/:id", authorize(Permission.DELETE_CATEGORIES), async (req, res) => {
-  if (isNaN(req.params.id)) {
-    res.status(400).json({ message: "Invalid ID" });
-  } else {
-    const category_id = await prisma.category.findFirst({ where: { id: Number(req.params.id) } });
-    if (!category_id) {
-      res.status(404).json({ message: "Category not found" });
+router.delete(
+  "/category/:id",
+  authorize(Permission.DELETE_CATEGORIES),
+  async (req, res) => {
+    if (isNaN(req.params.id)) {
+      res.status(400).json({ message: "Invalid ID" });
     } else {
-      await prisma.category.delete({
+      const category_id = await prisma.category.findFirst({
         where: { id: Number(req.params.id) },
       });
-      res.json({ message: "Category has been deleted" });
+      if (!category_id) {
+        res.status(404).json({ message: "Category not found" });
+      } else {
+        await prisma.category.delete({
+          where: { id: Number(req.params.id) },
+        });
+        res.json({ message: "Category has been deleted" });
+      }
     }
   }
-});
+);
 export default router;
